// src/app/models/dati-regioni-dto.ts
export class DatiRegioniDto {

  data?: string;
  stato?: string;
  denominazione_regione?: string;
  codice_regione?: number;
  ricoverati_conSintomi?: string;
  terapia_intensiva?: string;
  totale_ospedalizzati?: string;
  insolamento_domiciliare?: string;
  totale_attualmente_positivi?: string;
  nuovi_attualmente_positivi?: string;
  dimessi_guariti?: string;
  deceduti?: string;
  totale_casi?: number;
  tamponi?: string;


}